Restore double tap to wake string from CM 12.1.
The AOSP strings/translations of that feature are inconsistent to those
of the double tap to sleep feature, so replace the string to keep the
translations consistent.
As the same string exists in cm-12.1, all translations are already
present and will automatically be added in the next translation sync.

Change-Id: I613941b586fe60ea6b29e2695120359295289615

Conflicts:
	res/values/cm_strings.xml